Restructure MobileNav to use pages nav data and wrap categories list with suspense
